--- public/data/bestItems.json
@@ -1,70 +1,0 @@
-{
-  "bestItemData": [
-    {
-      "id": 1,
-      "categoryId": 1,
-      "productName": "Beef",
-      "price": 20000,
-      "weight": 300,
-      "productImgMain": "/images/kyle-mackie-MEnlQv-EQvY-unsplash.jpg",
-      "stock": 200,
-      "salesAmount": 100,
-      "isAntibioticFree": 0
-    },
-    {
-      "id": 2,
-      "categoryId": 2,
-      "productName": "Chicken",
-      "price": 5000,
-      "weight": 450,
-      "productImgMain": "/images/kyle-mackie-MEnlQv-EQvY-unsplash.jpg",
-      "stock": 300,
-      "salesAmount": 150,
-      "isAntibioticFree": 1
-    },
-    {
-      "id": 3,
-      "categoryId": 3,
-      "productName": "Pork",
-      "price": 9000,
-      "weight": 300,
-      "productImgMain": "/images/kyle-mackie-MEnlQv-EQvY-unsplash.jpg",
-      "stock": 250,
-      "salesAmount": 200,
-      "isAntibioticFree": 1
-    },
-    {
-      "id": 4,
-      "categoryId": 4,
-      "productName": "Beef",
-      "price": 20000,
-      "weight": 300,
-      "productImgMain": "/images/kyle-mackie-MEnlQv-EQvY-unsplash.jpg",
-      "stock": 200,
-      "salesAmount": 100,
-      "isAntibioticFree": 0
-    },
-    {
-      "id": 5,
-      "categoryId": 5,
-      "productName": "Chicken",
-      "price": 5000,
-      "weight": 450,
-      "productImgMain": "/images/kyle-mackie-MEnlQv-EQvY-unsplash.jpg",
-      "stock": 300,
-      "salesAmount": 150,
-      "isAntibioticFree": 0
-    },
-    {
-      "id": 6,
-      "categoryId": 6,
-      "productName": "Pork",
-      "price": 9000,
-      "weight": 300,
-      "productImgMain": "/images/kyle-mackie-MEnlQv-EQvY-unsplash.jpg",
-      "stock": 250,
-      "salesAmount": 200,
-      "isAntibioticFree": 1
-    }
-  ]
-}
