+++ src/components/BestItems/BestItemTitle/BestItemTitle.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './BestItemTitle.scss';
+
+export default function BestItemTitle() {
+  return (
+    <div>
+      <h5 className="bestItemTitle">저스트미트 베스트 상품</h5>
+    </div>
+  );
+}
