+++ src/components/BestItems/BestItemTitle/BestItemTitle.scss
@@ -1,0 +1,6 @@
+.bestItemTitle {
+  font-size: 32px;
+  line-height: 32px;
+  margin: 0px 0px 0px 0px;
+  font-family: 'Noto Sans KR', sans-serif;
+}
