--- src/components/List/BestItemList.scss
@@ -1,29 +1,0 @@
-.bestItemList {
-  margin-top: 80px;
-  padding: 0px 0px 0px 35px;
-  width: 100%;
-}
-
-.bestItemFlex {
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.bestItemTitle {
-  font-size: 32px;
-  line-height: 32px;
-  margin: 0px 0px 0px 0px;
-  font-family: 'Noto Sans KR', sans-serif;
-}
-
-.toShopWrapper {
-  text-align: center;
-  margin: 80px 0px;
-}
-
-.toShop {
-  width: 1100px;
-  height: 250px;
-  border-radius: 5px;
-  cursor: pointer;
-}
