--- src/components/List/BestItems.scss
@@ -1,57 +1,0 @@
-.bestItems {
-  margin-right: 20px;
-  padding: 0px 0px 0px 0px;
-}
-
-.productImg {
-  margin: 70px 0px 0px 40px;
-  width: 276px;
-  height: 276px;
-  border-radius: 5px;
-}
-
-.cartIcon {
-  width: 32px;
-  height: 32px;
-  cursor: pointer;
-}
-
-.cartIconWrapper {
-  height: 70px;
-  margin: 0px 10px 0px 0px;
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-}
-
-.bestItem {
-  margin: 32px 0px 0px;
-  width: 365px;
-  height: 416px;
-  background-color: #f9f7f8;
-  cursor: pointer;
-}
-
-.productInfo {
-  cursor: pointer;
-}
-
-.productName {
-  font-size: 18px;
-  width: 376px;
-  height: 27px;
-  margin: 16px 0px 0px 0px;
-  font-family: 'Noto Sans KR', sans-serif;
-}
-
-.productNum {
-  display: flex;
-  justify-content: space-between;
-  font-size: 30px;
-}
-
-.price {
-  font-size: 16px;
-  margin: 0px 0px 0px 0px;
-  font-family: 'Noto Sans KR', sans-serif;
-}
