--- src/components/List/Card.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import './Card.scss';
-
-const Card = props => {
-  const classes = 'card ' + props.className;
-  return <div className={classes}>{props.children}</div>;
-};
-
-export default Card;
