--- src/components/List/Card.scss
@@ -1,4 +1,0 @@
-.card {
-  // border-radius: 5px;
-  // box-shadow: 0 1px 8px rgba(0, 0, 0, 0.25);
-}
