+++ src/components/ToShop/ToShop.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import './ToShop.scss';
+
+export default function ToShop() {
+  return (
+    <div>
+      <Link to="/shop">
+        <div className="toShopWrapper">
+          <img
+            className="toShop"
+            alt="고기사진"
+            src="/images/kyle-mackie-qgfjZUXup1M-unsplash.jpg"
+          />
+        </div>
+      </Link>
+    </div>
+  );
+}
