+++ src/components/ToShop/ToShop.scss
@@ -1,0 +1,11 @@
+.toShop {
+  width: 1200px;
+  height: 300px;
+  margin: 80px 0px;
+  border-radius: 5px;
+}
+
+.toShopWrapper {
+  display: flex;
+  justify-content: center;
+}
