+++ src/pages/Home/Home.js
@@ -1,13 +1,13 @@
 import React from 'react';
-import BestItemList from '../../components/List/BestItemList'
+import BestItems from '../../components/BestItems/BestItems';
 import css from './Home.module.scss';
 
 function Home() {
   return (
     <div className={css.container}>
-      <BestItemList />
+      <BestItems />
     </div>
-  )
+  );
 }
 
 export default Home;
