--- src/pages/Item/ShopItems.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import ItemList from '../../components/List/ItemList';
-
-function ShopItems() {
-  return (
-    <div>
-      <ItemList />
-    </div>
-  );
-}
-
-export default ShopItems;
