+++ src/pages/Shop/Category/Category.js
@@ -1,13 +1,12 @@
 import React from 'react';
-import Card from './Card';
 import './Category.scss';
 
 const Category = ({ category }) => {
   return (
     <div className="CategoryWrapper">
-      <Card className="categories">
+      <div className="categories">
         <span>{category}</span>
-      </Card>
+      </div>
     </div>
   );
 };
