+++ src/pages/Shop/ItemCategory/ItemCategory.scss
@@ -1,0 +1,4 @@
+.itemCategory {
+  display: flex;
+  justify-content: flex-start;
+}
