+++ src/pages/Shop/ItemListBanner/ItemListBanner.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import './ItemListBanner.scss';
+
+export default function ItemListBanner() {
+  return (
+    <div>
+      <img
+        className="itemListBanner"
+        alt="고기사진"
+        src="/images/kyle-mackie-qgfjZUXup1M-unsplash.jpg"
+      />
+    </div>
+  );
+}
