+++ src/pages/Shop/ItemListBanner/ItemListBanner.scss
@@ -1,0 +1,4 @@
+.itemListBanner {
+  width: 100%;
+  height: 520px;
+}
